+++ example/lib/src/default_model_path_io.dart
@@ -1,5 +1,5 @@
 import 'dart:io';
 
 String defaultModelPath() {
-  return Platform.environment['FLUTTER_KIWI_FFI_MODEL_PATH'] ?? '';
+  return Platform.environment['FLUTTER_KIWI_NLP_MODEL_PATH'] ?? '';
 }