+++ lib/flutter_kiwi_ffi.dart
@@ -1,6 +1,1 @@
-export 'src/kiwi_analyzer_stub.dart'
-    if (dart.library.io) 'src/kiwi_analyzer_native.dart'
-    if (dart.library.js_interop) 'src/kiwi_analyzer_web.dart';
-export 'src/kiwi_exception.dart';
-export 'src/kiwi_options.dart';
-export 'src/kiwi_types.dart';
+export 'flutter_kiwi_nlp.dart';