+++ macos/Classes/flutter_kiwi_ffi.c
@@ -1,3 +1,3 @@
 // Relative import to be able to reuse the C sources.
-// See the comment in ../flutter_kiwi_ffi.podspec for more information.
+// See the comment in ../flutter_kiwi_nlp.podspec for more information.
 #include "../../src/flutter_kiwi_ffi.c"